requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.31.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
Flask==2.0.1
Flask-Login==0.6.3
fonttools==4.55.0
fsspec==2024.10.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==2.8
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.0.1
jiter==0.7.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
openai==1.54.4
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==2.0.0
